stopwatch class

+++ src/mikabooq.h
@@ -28,7 +28,7 @@ struct tcb_t {
 	state_t t_s; /* processor state */
 
 	int t_status;
-	unsigned int run_time; //milliseconds of CPU time used. check scheduler.c for the accounting function
+	uint64_t run_time; //milliseconds of CPU time used. check scheduler.c for the accounting function
 	int errno; /* error status of the thread */
 
 	struct tcb_t *t_wait4sender; /* expected sender (if t_status == T_STATUS_W4MSG), NULL means accept msg from anybody */
